requirements.txt
streamlit>=1.28.0
google-genai>=1.0.0
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
python-dotenv>=1.0.0
opencv-python>=4.8.0
sounddevice>=0.4.6
numpy>=1.24.0
groq>=1.0.0
